@@ -1,6 +1,6 @@
 module github.com/mg52/go-api
 
-go 1.18
+go 1.21
 
 require (
 	github.com/go-playground/validator/v10 v10.11.1
